% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggman.R
\name{check.input.ggman}
\alias{check.input.ggman}
\title{This function checks the user input for the function ggman.R}
\usage{
check.input.ggman(gwas = gwas, clumps = clumps, snp = snp, bp = bp,
  chrom = chrom, pvalue = pvalue, sigLine = sigLine,
  lineColour = lineColour, pointSize = pointSize, ymin = ymin,
  ymax = ymax, logTransform = logTransform, invert = invert,
  invert.method = invert.method, invert.var = invert.var,
  relative.positions = relative.positions, xlabel = xlabel,
  ylabel = ylabel, title = title, legend.title = legend.title,
  clumps.label.type = clumps.label.type, legend.remove = legend.remove)
}
\description{
This function checks the user input for the function ggman.R
}
\keyword{internal}

