% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggmanClumps.R
\name{check.input.ggmanClumps}
\alias{check.input.ggmanClumps}
\title{This function checks the user input for the function ggmanClumps.R}
\usage{
check.input.ggmanClumps(plink.clumped = plink.clumped,
  index.snp.column = index.snp.column, clumps.column = clumps.column,
  label.column = label.column, group.column = group.column)
}
\description{
This function checks the user input for the function ggmanClumps.R
}
\keyword{internal}

