% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggmanHighlightGroup.R
\name{check.input.ggmanHighlightGroup}
\alias{check.input.ggmanHighlightGroup}
\title{This function checks the input for the function ggmanHighlightGroup}
\usage{
check.input.ggmanHighlightGroup()
}
\description{
This function checks the input for the function ggmanHighlightGroup
}
\keyword{internal}

