% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggmanZoom.R
\name{check.input.ggmanZoom}
\alias{check.input.ggmanZoom}
\title{This function checks the user input for the function ggmanZoom}
\usage{
check.input.ggmanZoom()
}
\description{
This function checks the user input for the function ggmanZoom
}
\keyword{internal}

