% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggchicago.R
\name{ggchicago}
\alias{ggchicago}
\title{Plot a chicago plot using ggplot2.}
\usage{
ggchicago(data1, data2, chr = "CHR", bp = "BP", P = "P", logP = TRUE,
  build = "hg19", significance = c(5e-08),
  theme_base = theme_publication(), scale_color = scale_color_traditional())
}
\arguments{
\item{chr}{}

\item{bp}{}

\item{P}{}

\item{logP}{}

\item{build}{}

\item{theme_base}{}

\item{scale_color}{}

\item{data}{}
}
\description{
Plot a chicago plot using ggplot2.
}

