% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggclumps.R
\name{ggClumps}
\alias{ggClumps}
\title{ggClumps}
\usage{
ggClumps(plink.clumped)
}
\arguments{
\item{plink.clumped}{A data frame read from plink.clumped file.}
}
\description{
Process the clumped file from plink
}
\examples{

}

