% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmanhattan.R
\name{ggmanhattan}
\alias{ggmanhattan}
\title{Plot a manhattan plot using ggplot2.}
\usage{
ggmanhattan(data, chr = "CHR", bp = "BP", P = "P", logP = TRUE,
  build = "hg19", theme_base = theme_publication(),
  scale_color = scale_color_traditional())
}
\arguments{
\item{data}{}

\item{chr}{}

\item{bp}{}

\item{P}{}

\item{logP}{}

\item{build}{}

\item{theme_base}{}

\item{scale_color}{}
}
\description{
Plot a manhattan plot using ggplot2.
}

