% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanInvert.R
\name{ggmanInvert}
\alias{ggmanInvert}
\title{A ggPlot wrapper for Manhattan Plot}
\usage{
ggmanInvert(gwas, snp = "snp", bp = "bp", chrom = "chrom",
  pvalue = "pvalue", effect = "or", method = "or", sigLine = 8,
  lineColor = "red", pointSize = 0.1, ymin = NA, ymax = NA,
  xlabel = "chromosome", ylabel = "-log10(P value)",
  title = "Manhattan Plot")
}
\arguments{
\item{gwas}{A data frame object with the gwas results}

\item{snp}{Name of the column containing the markers to plot}

\item{bp}{Name of the column containing the basepair positions}

\item{chrom}{Name of the column containing the chromosome information}

\item{pvalue}{Name of the column containing the p values to plot}

\item{sigLine}{Threshold for the genome wide significant line in -log10 scale}

\item{lineColor}{colour of the genomewide line}

\item{pointSize}{Size of the points in the plot supplied to geom_point}

\item{ymin}{Starting point of y axis}

\item{ymax}{Ending point of y axis}

\item{xlabel}{x axis label}

\item{ylabel}{y  axis label}

\item{title}{plot title}

\item{logTransform}{Whether to log transform the p values}
}
\value{
If unassigned to a variable, returns a manhattan plot; If assigned to a variable, stored as a ggplot layer
}
\description{
Creates a ggPlot layer for Manhattan Plot
}
\examples{

}

