% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanLabel.R
\name{ggmanLabel}
\alias{ggmanLabel}
\title{Add labels to ggman Plot}
\usage{
ggmanLabel(ggmanPlot, labelDfm, snp = "snp", label = "label",
  type = "label", colour = "black", size = 2)
}
\arguments{
\item{ggmanPlot}{A ggman plot object}

\item{labelDfm}{A data frame containing the snps and the labels}

\item{snp}{Name of the column containing the markers;}

\item{label}{Name of the column containing the labels}

\item{type}{type of label; either "label" or "text"}

\item{colour}{Colour of the label}

\item{size}{of the label}
}
\description{
Adds user provided labels to the ggman Manhattan plot
}
\examples{

}

