% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanZoom.refseq.R
\name{ggmanZoom.refseq}
\alias{ggmanZoom.refseq}
\title{ggmanZoom.refseq}
\usage{
ggmanZoom.refseq(ggmanPlot, chromosome, start.position = NA,
  end.position = NA, xlabel = NA, ylabel = NA, title = NA,
  highlight.group = NA, legend.title = "legend", legend.remove = FALSE,
  ymax = 10, ymin = 0, genome = "hg19", exon.width = 0.5,
  gene.width = 0.05, ...)
}
\arguments{
\item{ggmanPlot}{A ggplot object of class 'ggman'}

\item{chromosome}{Chromosome identifier}

\item{start.position}{Starting basepair position}

\item{end.position}{Ending basepair position}

\item{xlabel}{X axis label}

\item{ylabel}{Y axis label}

\item{title}{Plot title}

\item{highlight.group}{Name of the column containing grouping variable in the parent data.frame that was used to create main ggplot layer.}

\item{legend.title}{title for legend; default is 'legend'}

\item{legend.remove}{if TRUE, the legend will be removed; default is FALSE}

\item{...}{other arguments to pass to \code{\link{geom_point}}}
}
\value{
A regional association plot
}
\description{
Zoom in to a specific region of the Manhattan Plot
}
\examples{

#specific chromosome
p1 <- ggman(toy.gwas, snp = "snp", bp = "bp", chrom = "chrom",
pvalue = "pvalue")
ggmanZoom(p1, chromosome = 1)

#specific region
ggmanZoom(p1, chromosome = 1, start.position = 215388741, end.position = 238580695)

#add highlights and legend
ggmanZoom(p1, chromosome = 1, start.position = 215388741, end.position = 238580695,
          highlight.group = "gene", legend.title = "Genes")


}

