% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\docType{data}
\name{gwas}
\alias{gwas}
\title{A simulated GWAS dataset}
\format{A dataframe with simulated GWAS results (simulated using plink 1.9)}
\usage{
data(gwas)
}
\description{
GWAS data simulated using plink 1.9 --simulate option.
}
\examples{
data(gwas)
p1 <- ggman(gwas, snp = "SNP", chrom = "CHR", bp = "BP", pvalue = "P")

gwas.sig <- gwas[-log10(gwas$P)>8,]

ggmanLabel(p1, labelDfm = gwas.sig, snp = "SNP", label = "SNP")

}
\keyword{datasets}

