% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlights.R
\docType{data}
\name{highlights}
\alias{highlights}
\title{highlights}
\format{A vector of snp IDs}
\usage{
data(highlights)
}
\description{
A vector of markers to highlight
}
\examples{
data(highlights)

p1 <- ggman(gwas, snp = "SNP", chrom = "CHR", bp = "BP", pvalue = "P")

ggmanHighlight(p1, highlight = highlights, colour = "red")

}
\keyword{datasets}

