% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scale-colour-traditional.R
\name{scale_colour_traditional}
\alias{scale_colour_traditional}
\title{custom ggplot2 scale_colour for a manhattan plot}
\usage{
scale_colour_traditional()
}
\description{
custom ggplot2 scale_colour for a manhattan plot
}

