% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme-publication.R
\name{theme_publication}
\alias{theme_publication}
\title{custom ggplot2 theme for publication}
\usage{
theme_publication(base_size = 16, base_family = "Helvetica", ...)
}
\arguments{
\item{base_size}{}

\item{base_family}{}

\item{...}{further arguments passed to \code{theme_classic}.}
}
\description{
custom ggplot2 theme for publication
}

