% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.clumped.R
\docType{data}
\name{toy.clumped}
\alias{toy.clumped}
\title{Randomly generated SNP clumps}
\format{A data frame containg SNP clumps in the format similar to plink \code{--clump} output.}
\usage{
data(toy.clumped)
}
\description{
The dataset 'toy.clumped' is a data.frame object containing the index SNPs and the SNP clumps.
}
\details{
The file is similar to a 'plink.clumped' output of plink. The plink.clumped file read into a data.frame object and only the columns, "SNP" and "SP2" are retained.
}
\keyword{datasets}

