% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.gwas.R
\docType{data}
\name{toy.gwas}
\alias{toy.gwas}
\title{Randomly generated toy GWAS dataset}
\format{A dataframe with simulated GWAS results (simulated using plink 1.9)}
\usage{
data(toy.gwas)
}
\value{
A data frame
}
\description{
The dataset 'toy.gwas' is a data.frame object with the summary results of a GWAS analysis.
}
\keyword{datasets}

