% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.highlights.group.R
\docType{data}
\name{toy.highlights.group}
\alias{toy.highlights.group}
\title{toy.highlights.group}
\format{A dataframe}
\usage{
data(toy.highlights.group)
}
\description{
A data.frame containing the SNPs, labels and grouping variable.
}
\keyword{datasets}

