% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.highlights.R
\docType{data}
\name{toy.highlights}
\alias{toy.highlights}
\title{toy.highlights}
\format{A vector of snp IDs}
\usage{
data(toy.highlights)
}
\description{
A vector of markers to highlight
}
\keyword{datasets}

