% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggman.R
\name{check.input.ggman}
\alias{check.input.ggman}
\title{This function checks the user input for the function ggman.R}
\usage{
check.input.ggman()
}
\value{
Nothing; internal function.
}
\description{
This function checks the user input for the function ggman.R
}
\keyword{internal}

