% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggmanHighlight.R
\name{check.input.ggmanHighlight}
\alias{check.input.ggmanHighlight}
\title{This function checks the input arguments to the function ggmanHighlight}
\usage{
check.input.ggmanHighlight()
}
\value{
Nothing; internal function.
}
\description{
This function checks the input arguments to the function ggmanHighlight
}
\keyword{internal}

