% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.input.ggmanLabel.R
\name{check.input.ggmanLabel}
\alias{check.input.ggmanLabel}
\title{This function checks the user input for the function ggmanLabel.R}
\usage{
check.input.ggmanLabel(ggmanPlot = ggmanPlot, labelDfm = labelDfm,
  snp = snp, label = label, type = type)
}
\description{
This function checks the user input for the function ggmanLabel.R
}
\keyword{internal}

