% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggman.R
\name{ggman}
\alias{ggman}
\title{Creates a Manhattan Plot.}
\usage{
ggman(gwas, clumps = NA, snp = "snp", bp = "bp", chrom = "chrom",
  pvalue = "pvalue", sigLine = 8, lineColour = "red", pointSize = 0.1,
  ymin = 0, ymax = NA, logTransform = TRUE, relative.positions = FALSE,
  xlabel = "chromosome", ylabel = "-log10(P value)",
  title = "Manhattan Plot", legend.title = "legend",
  clumps.label.type = "label", legend.remove = FALSE, ...)
}
\arguments{
\item{gwas}{A data frame with the gwas results}

\item{clumps}{Optional argument; takes an object of class 'ggclumps' containing
the SNP clumps, see \code{\link{ggmanClumps}}}

\item{snp}{Name of the column containing SNP identifiers; default is 'snp'}

\item{bp}{Name of the column containing the basepair positions; default is 'bp'}

\item{chrom}{Name of the column containing the chromosome identifers; default is 'chrom'}

\item{pvalue}{Name of the column containing the p values; default is 'pvalue'}

\item{sigLine}{Threshold for the genome wide significant line in -log10 scale; default is 8; specify NA to remove the line.}

\item{lineColour}{colour of the genomewide line; default is 'red'}

\item{pointSize}{Size of the points in the plot; default is 0.1}

\item{ymin}{Starting point of y axis; default is 0}

\item{ymax}{Ending point of y axis}

\item{logTransform}{if TRUE, P value is -log10 transformed; default is TRUE; Specify FALSE
when plotting values other p values, such as zscore or beta}

\item{relative.positions}{if TRUE, the x axis points will be calculated in proportion to the basepair positions. So, the gaps in the genome with no genotypes will be reflected in the plot(requires more computation, hence more time to plot). If FALSE, the SNPs are ascendingly sorted chromosome-wise. The default value is FALSE.}

\item{xlabel}{X-axis label}

\item{ylabel}{Y-axis label}

\item{title}{plot title}

\item{legend.title}{title of legend; this argument applies only when the clumps are plotted and highlighted in groups.}

\item{clumps.label.type}{type of label; either 'text' or 'label'; default is 'label'}

\item{legend.remove}{if TRUE, the  legend will be removed; default is FALSE}

\item{...}{other arguments to pass to \code{\link[ggplot2]{geom_point}};
Note: do not use \emph{size} to specify size of the points, instead use \emph{pointSize}}
}
\description{
Creates a Manhattan Plot as a ggplot layer.
}
\details{
The function creates a manhattan plot as a ggplot layer. The output can be stored in a variable,
and additional layers can be added. See \code{\link{ggmanHighlight}}, \code{\link{ggmanZoom}} and
 \code{\link{ggmanLabel}}.
}
\examples{

#basic plot
ggman(gwas,snp = "SNP", bp ="BP", chrom = "CHR", pvalue = "P")


\dontrun {
#with relative positions
ggman(gwas,snp = "SNP", bp ="BP", chrom = "CHR", pvalue = "P", relative.positions = TRUE)
#with clumps
ggman(gwas,snp = "SNP", bp ="BP", chrom = "CHR", pvalue = "P", clumps = gwasclumps)
}

}

