% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanClumps.R
\name{ggmanClumps}
\alias{ggmanClumps}
\title{ggmanClumps}
\usage{
ggmanClumps(plink.clumped, index.snp.column, clumps.column, label.column = NA,
  group.column = NA)
}
\arguments{
\item{plink.clumped}{A data frame read from plink.clumped file.}
}
\description{
Process the clumped file from plink
}
\examples{

}

