% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanHighlight.R
\name{ggmanHighlight}
\alias{ggmanHighlight}
\title{Highlight points in Manhattan plot}
\usage{
ggmanHighlight(ggmanPlot, highlight, colour = "red", size = 0.1, ...)
}
\arguments{
\item{ggmanPlot}{A ggman plot}

\item{highlight}{A character vector of SNP ids to highlight}

\item{colour}{highlight colour}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_pont}}}
}
\value{
A manhattan plot with highlighted markers
}
\description{
Highlights a specific subset of points in the ggman plot
}
\examples{

}

