% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanHighlight.R
\name{ggmanHighlight}
\alias{ggmanHighlight}
\title{Highlight markers in Manhattan plot}
\usage{
ggmanHighlight(ggmanPlot, highlight, colour = "green", pointSize = 0.1)
}
\arguments{
\item{ggmanPlot}{A ggman plot}

\item{highlight}{A character vector of SNP ids to highlight}

\item{colour}{Colour of the highlight}

\item{pointSize}{Size of the highlight points}
}
\value{
A manhattan plot with highlighted markers
}
\description{
Highlights a specific subset of markers in the ggman plot
}
\examples{

}

