% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanHighlight.R
\name{ggmanHighlight}
\alias{ggmanHighlight}
\title{Highlight points in Manhattan plot}
\usage{
ggmanHighlight(ggmanPlot, highlight, colour = "red", ...)
}
\arguments{
\item{ggmanPlot}{A ggman plot of class 'ggman'}

\item{highlight}{A character vector of SNP ids to highlight}

\item{colour}{highlight colour}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_point}}}
}
\value{
A manhattan plot with highlighted markers
}
\description{
Highlights a specific subset of points in the Manhattan plot
}
\examples{

ggmanHighlight(p1, highlight = toy.highlights)

}

