% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanHighlightGroup.R
\name{ggmanHighlightGroup}
\alias{ggmanHighlightGroup}
\title{Highlight groups of points in Manhattan plot}
\usage{
ggmanHighlightGroup(ggmanPlot, highlightDfm, snp = "snp", group = "group",
  legend.title = "legend", legend.remove = FALSE, ...)
}
\arguments{
\item{ggmanPlot}{A ggman plot of class 'ggman'; see \code{\link{ggman}}}

\item{highlightDfm}{A data frame object; one of the columns should contain snps identifiers}

\item{snp}{Name of the snp column}

\item{group}{Name of the grouping column; if all the snps are to be highlighted with same colour use \code{\link{ggmanHighlight}}}

\item{legend.title}{Title of the legend.}

\item{legend.remove}{If TRUE, legend will be removed.}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_point}}}
}
\value{
A manhattan plot with highlighted markers
}
\description{
Highlights groups of points in the ggman Manhattan Plot and add a legend.
}
\examples{

}

