% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanLabel.R
\name{ggmanLabel}
\alias{ggmanLabel}
\title{Add labels to ggman Plot}
\usage{
ggmanLabel(ggmanPlot, labelDfm, snp = "snp", label = "label",
  type = "label", ...)
}
\arguments{
\item{ggmanPlot}{A ggplot layer of class 'ggman'; see \code{\link{ggman}}}

\item{labelDfm}{A data frame object with the SNPs and the labels;}

\item{snp}{Name of the column containing the markers;}

\item{label}{Name of the column containing the labels; It can be the snp column itself.}

\item{type}{type of label; either "label" or "text"; see details section}

\item{...}{other arguments applicable to \code{\link[ggrepel]{geom_text_repel}} or \code{\link[ggrepel]{geom_label_repel}}}
}
\description{
Label points in the ggman plot.
}
\details{
This function adds a layer of textual annotation using \code{\link[ggrepel]{geom_label_repel }}(if type = "label") or \code{\link[ggrepel]{geom_text_repel}} (if type = "text")
}
\examples{


}

