% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanZoom.R
\name{ggmanZoom}
\alias{ggmanZoom}
\title{ggmanZoom}
\usage{
ggmanZoom(ggmanPlot, chromosome, start.position = NA, end.position = NA,
  gene.tracks = TRUE, ymax = 10, genome = "hg19", exon.width = 0.5,
  gene.width = 0.05, stack.level = 1, remove.gene.text.overlap = FALSE,
  gene.text.size = 2, xlabel = NA, ylabel = NA, title = NA, ...)
}
\arguments{
\item{ggmanPlot}{A ggplot object of class 'ggman'}

\item{chromosome}{Chromosome identifier}

\item{start.position}{Starting basepair position}

\item{end.position}{Ending basepair position}

\item{gene.tracks}{if TRUE, refseq gene tracks will be downloaded frm UCSC server and plotted. default is TRUE}

\item{ymax}{maximum limit of y axis}

\item{genome}{genome build version; default is 'hg19'}

\item{exon.width}{width of the exons in the gene tracts; default is 0.5}

\item{gene.width}{with of the introns in the gene tracts; default is 0.05}

\item{stack.level}{levels of stacking of the gene tracts; default is 1, where all the genes are plotted in single track}

\item{remove.gene.text.overlap}{if TRUE, one of the overlapping gene text labels will be removed}

\item{gene.text.size}{size of gene labes below the gene tracts; default is 2}

\item{xlabel}{X axis label}

\item{ylabel}{Y axis label}

\item{title}{Plot title}

\item{...}{other arguments to pass to \code{\link{geom_point}}}
}
\value{
A regional association plot
}
\description{
Zoom in to a specific region of the Manhattan Plot
}
\examples{

#specific chromosome
p1 <- ggman(toy.gwas, snp = "snp", bp = "bp", chrom = "chrom",
pvalue = "pvalue")
ggmanZoom(p1, chromosome = 1)

#specific region
ggmanZoom(p1, chromosome = 1, start.position = 215388741, end.position = 238580695)





}

