% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanZoom.R
\name{ggmanZoom}
\alias{ggmanZoom}
\title{ggmanZoom}
\usage{
ggmanZoom(ggmanPlot, chromosome, start.position, end.position, xlabel = NA,
  ylabel = NA, title = NA, ...)
}
\arguments{
\item{ggmanPlot}{A ggplot object of class 'ggplot'}

\item{chromosome}{Chromosome identifier}

\item{start.position}{Starting basepair position}

\item{end.position}{Ending basepair position}

\item{xlabel}{X axis label}

\item{ylabel}{Y axis label}

\item{title}{Plot title}

\item{...}{other arguments to pass to \code{\link{geom_point}}}
}
\description{
Zoom in to a specific region of the Manhattan Plot
}
\examples{


}

